[chore] Animator 파일 분리
