+++ NaverMapA/NaverMapA/Model/Animator/Animatorable.swift
@@ -1,0 +1,12 @@
+//
+//  Animatorable.swift
+//  NaverMapA
+//
+//  Created by 홍경표 on 2020/12/09.
+//
+
+import Foundation
+
+protocol Animatorable: Operation {
+    func markerAnimation()
+}
